Initial work on shifting to manifest V3
It's kind of a mess at the moment, and very much a WIP, but the popup works and displays markdown!
It's looking more and more like I'll just have to ditch the background worker/page entirely and do more in the popup itself.
I also added the popup as a sidebar thing! Surprisingly, this works least well in Firefox: the sidebar doesn't have host permission for the active tab.
I still have to work out downloading, as well as all the context menu options and such, but we'll see how we go...