Got downloads working from the popup script
Also, got downloading images working without the downloads API — remains to be seen if that works on Safari, though.
Still some issues with the naming of images... Subfolders don't work without the downloads API and there's something funny going on with the url encoding, so even though it appears to download all the files correctly and correctly populate the (relative) links, the images still won't show up in the markdown file.